fix FTBS on 64bit Linux
On 64bit the long int type is 64 bit, but the MSR operations still
are 32bit, thus we need to use uint32_t. Also need slighty modified
versions of the inline asm code, since on 64bit machines, the 32bit
(pseudo-)registers can't be used on push/pop.

This patch is mostly for CI purpose, since there isn't any known version
of this graphics device on a 64 bit machine.

+++ src/cim/cim_parm.h
@@ -113,9 +113,8 @@ typedef struct tagGeodeLinkNode {
 /*---------------------------------------------*/
 
 typedef struct tagQ_WORD {
-    unsigned long high;
-    unsigned long low;
-
+    uint32_t high;
+    uint32_t low;
 } Q_WORD;
 
 /*===================================================*/
